microservices/user-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
prometheus-flask-exporter==0.23.0